backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
Pillow==11.1.0
python-dotenv==1.0.1
